automatically generate encryption SEED key with 5~8 arrays

+++ tools/autogen.sh
@@ -3,6 +3,14 @@
 export LIBTOOL=/usr/bin/libtool
 export AUTOMAKE=/usr/bin/automake
 export ACLOCAL=/usr/bin/aclocal
+
+# chekc config.h of mod_screwim
+if [ ! -f ../config.h ]; then
+	echo "ERROR: Not found ../config.h" >& /dev/stderr
+	echo "       First, move location to \"../\" and run \"./configure\""
+	exit 1
+fi
+
 rm -f configure autotool/{config.guess,config.sub,ltmain.sh}
 autoreconf --install
 rm -rf autom4te* *~
